added TRS MIDI output
I added a TRS-A MIDI output to my unit for using with my Dirtywave M8 when it shows up in a few months. This was accomplished by designing an expanded housing (5mm taller), mounting a 3.5mm jack inside the unit and then wiring the pin to the TX port of the ESP32 (with a 10ohm resistor) and the sleeve to the 3.3V line (with a 33ohm resistor), ground to ground.