revision 2 of TRS housing
This is a remake of the housing to improve ergonomics (the sharp edges at the bottom bothered my hands) as well as add a bit more room inside for a larger battery.

The standard rev2.stl has a hole in the side to mimic the original housing and its charging insert/window. You can print the -glowplug.stl in a clear filament to insert into the hole and carry the light from the charging board.

If using a lighter coloured/transparent/transulecent filament you can just print the -nowindow version of the housing as it has no hole in the side (not needed when you can see the charging light through the housing walls).